+++ django_sendgrid_webhook/tests/test_signals.py
@@ -1,6 +1,6 @@
 from django.test import TestCase, Client
 
-from django_sendgrid import utils, signals
+from django_sendgrid_webhook import utils, signals
 
 import json
